75 launcher io layers (#83)
* Some AdvantageKit stuff

Also autos is broken. Too bad!

* It will be rewritten

trust

* Update Robot.kt

* LauncherIO interface and inputs

* package

* Better IO layer to match Falon's conventions

* roller pid and ff methods

* literally 4 characters

unless you count spaces. then it's 6

* First attempt at impl, yell at me when its wrong

* Separation of flywheels

* Finished Separating flywheels

* Rotational Inertia Units

Trust me this is related to io layers

* Probably decent

* checkup

* Feedforward added

This should be good, except for docs

* Roller position, flywheel linear control

I am confused as to what is meant by "roller linear position control," as well as why you requested roller position be logged when it already is…

---------

Co-authored-by: Europa <falon.n.clark@gmail.com>